# openapi.yaml
openapi: 3.0.3
info:
  title: SiteTrace Website Intelligence API Suite
  version: 1.0.0
  description: >
    Ten website intelligence endpoints for SEO, crawlability, social previews,
    structured data, contact extraction, tech stack detection, security headers,
    lead scoring, and full website reports.
  contact:
    name: SiteTrace
servers:
  - url: https://api-marketplace-b73f.onrender.com
    description: Production Render service
security:
  - RapidApiKey: []
...
paths:
...
  /api/security-headers:
    post:
      summary: Security Headers
      description: Audit HTTP security headers and SSL metadata, returning a security score, grade, missing headers, and recommendations by plan.
      operationId: securityHeaders
      tags:
        - Security
      parameters:
        - $ref: "#/components/parameters/RapidApiSubscription"
        - $ref: "#/components/parameters/SiteTracePlan"
      requestBody:
        $ref: "#/components/requestBodies/UrlRequest"
      responses:
        "200":
          description: Security header audit result.
          headers:
            X-RateLimit-Plan:
              $ref: "#/components/headers/XRateLimitPlan"
            X-RateLimit-Limit:
              $ref: "#/components/headers/XRateLimitLimit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/XRateLimitRemaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/XRateLimitReset"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseAnalysisResponse"
                  - type: object
                    properties:
                      security_score:
                        type: integer
                      grade:
                        type: string
                      headers_present:
                        type: integer
                      headers_missing:
                        type: integer
                      ssl:
                        type: object
                        nullable: true
                      headers:
                        type: array
                        items:
                          type: object
              example:
                success: true
                url: https://example.com/
                status_code: 200
                grade: C
                headers_present: 2
                headers_missing: 7
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/RateLimited"
        "502":
          $ref: "#/components/responses/UpstreamFailure"
...
  /api/report-generator:
    post:
      summary: Report Generator
      description: Generate a combined website audit report covering SEO, social metadata, performance, security, content, and crawlability signals.
      operationId: reportGenerator
      tags:
        - Reports
        - SEO
        - Security
      parameters:
        - $ref: "#/components/parameters/RapidApiSubscription"
        - $ref: "#/components/parameters/SiteTracePlan"
      requestBody:
        $ref: "#/components/requestBodies/UrlRequest"
      responses:
        "200":
          description: Full report result.
          headers:
            X-RateLimit-Plan:
              $ref: "#/components/headers/XRateLimitPlan"
            X-RateLimit-Limit:
              $ref: "#/components/headers/XRateLimitLimit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/XRateLimitRemaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/XRateLimitReset"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseDomainResponse"
                  - type: object
                    properties:
                      overall_score:
                        type: integer
                      overall_grade:
                        type: string
                      scores:
                        type: object
                        additionalProperties: true
                      seo:
                        type: object
                        additionalProperties: true
                      social:
                        type: object
                        additionalProperties: true
                      performance:
                        type: object
                        additionalProperties: true
                      security:
                        type: object
                        additionalProperties: true
                      crawlability:
                        type: object
                        additionalProperties: true
                      insights:
                        $ref: "#/components/schemas/AuditInsights"
              example:
                success: true
                url: https://example.com/
                domain: example.com
                status_code: 200
                overall_score: 80
                overall_grade: B
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/RateLimited"
        "502":
          $ref: "#/components/responses/UpstreamFailure"
components:
  securitySchemes:
    RapidApiKey:
      type: apiKey
      in: header
      name: X-RapidAPI-Key
      description: RapidAPI gateway key. Direct production access also requires the configured RapidAPI proxy secret.
  parameters:
    RapidApiSubscription:
      name: X-RapidAPI-Subscription
      in: header
      required: false
      description: RapidAPI plan header. BASIC maps to free, PRO maps to pro, ULTRA maps to ultra, and MEGA maps to mega.
      schema:
        type: string
        enum:
          - BASIC
          - PRO
          - ULTRA
          - MEGA
      example: PRO
    SiteTracePlan:
      name: X-Plan
      in: header
      required: false
      description: Direct/internal plan override. If omitted, the API defaults to free unless X-RapidAPI-Subscription is present.
      schema:
        type: string
        enum:
          - free
          - pro
          - ultra
          - mega
      example: pro
  headers:
    XRateLimitPlan:
      description: Effective plan used by the API.
      schema:
        type: string
        enum:
          - free
          - pro
          - ultra
          - mega
    XRateLimitLimit:
      description: Maximum requests allowed in the current one-minute window.
      schema:
        type: integer
    XRateLimitRemaining:
      description: Remaining requests in the current one-minute window.
      schema:
        type: integer
    XRateLimitReset:
      description: Unix timestamp when the current rate-limit window resets.
      schema:
        type: integer
  requestBodies:
    UrlRequest:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/UrlRequest"
          example:
            url: https://example.com
  responses:
    BadRequest:
      description: Invalid request, missing URL, unsupported protocol, or private/local URL.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            error: Private/local network URLs are not supported
    RateLimited:
      description: The request exceeded the active plan's per-minute rate limit.
      headers:
        Retry-After:
          description: Seconds until the client should retry.
          schema:
            type: integer
        X-RateLimit-Plan:
          $ref: "#/components/headers/XRateLimitPlan"
        X-RateLimit-Limit:
          $ref: "#/components/headers/XRateLimitLimit"
        X-RateLimit-Remaining:
          $ref: "#/components/headers/XRateLimitRemaining"
        X-RateLimit-Reset:
          $ref: "#/components/headers/XRateLimitReset"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/RateLimitResponse"
          example:
            success: false
            error: Rate limit exceeded
            plan: free
            limit: 5
            window_seconds: 60
            retry_after_seconds: 60
    UpstreamFailure:
      description: The target website could not be fetched or analyzed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            error: Failed to fetch or analyze URL
...
  schemas:
    UrlRequest:
      type: object
      required:
        - url
      properties:
        url:
          type: string
          format: uri
          description: Public HTTP or HTTPS website URL to analyze.
          example: https://example.com
        max_urls:
          type: integer
          description: Optional sitemap result limit for /api/sitemap-analyzer.
          minimum: 1
          maximum: 2000
          example: 500
...
    BaseAnalysisResponse:
      type: object
      required:
        - success
        - url
      properties:
        success:
          type: boolean
        url:
          type: string
          format: uri
        status_code:
          type: integer
        analyzed_at:
          type: string
          format: date-time
        final_url:
          type: string
          format: uri
    BaseDomainResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
        url:
          type: string
          format: uri
        domain:
          type: string
        status_code:
          type: integer
        analyzed_at:
          type: string
          format: date-time
    AuditInsights:
      type: object
      properties:
        executive_summary:
          type: string
        page_profile:
          type: object
          additionalProperties: true
        top_priorities:
          type: array
          items:
            type: object
            properties:
              priority:
                type: string
              area:
                type: string
              issue:
                type: string
              action:
                type: string
        quick_wins:
          type: array
          items:
            type: string
        content_opportunities:
          type: array
          items:
            type: object
            properties:
              priority:
                type: string
              title:
                type: string
              primary_keyword:
                type: string
              supporting_keywords:
                type: array
                items:
                  type: string
              intent:
                type: string
        optional_content_plan:
          type: object
          additionalProperties: true
    ErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
        detail:
          type: string
    RateLimitResponse:
      type: object
      required:
        - success
        - error
        - plan
        - limit
        - window_seconds
        - retry_after_seconds
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: Rate limit exceeded
        plan:
          type: string
          enum:
            - free
            - pro
            - ultra
            - mega
        limit:
          type: integer
        window_seconds:
          type: integer
          example: 60
        retry_after_seconds:
          type: integer